jobs/requirements.txt
azure-storage-blob
requests
pillow
python-dotenv
